+++ translator/src/main/scala/soda/translator/replacement/Replacement.scala
@@ -43,8 +43,7 @@ case class Replacement (line: String ) {
     Replacement (replace_all (line, pattern, replacement )  )
 
   def replace_all (line: String, pattern: String, replacement: String ): String = {
-    lazy val result = postproc (
-      Rec () .foldLeftWhile (Rec () .range (line.length ), initial_value, next_value, condition )
+    lazy val result = postproc (Rec () .foldLeftWhile (Rec () .range (line.length ), initial_value, next_value, condition )
     )
 
     lazy val initial_value = FoldTuple (Seq (), 0 )