+++ translator/src/test/scala/soda/lib/SeqSDSpec.scala
@@ -7,8 +7,7 @@ case class SeqSDSpec () extends AnyFunSuite {
   test ("should detect an empty sequence") {
     lazy val input = Seq ()
     lazy val seqsd = SeqSDBuilder () .build (input )
-    lazy val obtained = seqsd.open (
-      ifEmpty = true, ifNonEmpty = nonEmpty => false
+    lazy val obtained = seqsd.open (ifEmpty = true, ifNonEmpty = nonEmpty => false
     )
     assert (obtained )
   }
@@ -16,8 +15,7 @@ case class SeqSDSpec () extends AnyFunSuite {
   test ("should detect an non empty sequence") {
     lazy val input = Seq (1 )
     lazy val seqsd = SeqSDBuilder () .build (input )
-    lazy val obtained = seqsd.open (
-      ifEmpty = false, ifNonEmpty = nonEmpty => true
+    lazy val obtained = seqsd.open (ifEmpty = false, ifNonEmpty = nonEmpty => true
     )
     assert (obtained )
   }
@@ -31,8 +29,7 @@ case class SeqSDSpec () extends AnyFunSuite {
     lazy val input = Seq (2, 7, 1, 8, 2, 8, 1, 8, 2, 8, 4, 5, 9 )
     lazy val expected = SomeSD [Int]  (9 )
     lazy val obtained =
-      SeqSDBuilder () .build (input ) .open (
-        ifEmpty = NoneSD (), ifNonEmpty = sequence => SomeSD (max (sequence )  )
+      SeqSDBuilder () .build (input ) .open (ifEmpty = NoneSD (), ifNonEmpty = sequence => SomeSD (max (sequence )  )
       )
     assert (obtained == expected )
   }
@@ -47,8 +44,7 @@ case class SeqSDSpec () extends AnyFunSuite {
   test ("should reverse another sequence") {
     lazy val input = SeqSDBuilder () .build (Seq (2, 7, 1, 8, 2, 8, 1, 8, 2, 8, 4, 5, 9 ) )
     lazy val expected = SeqSDBuilder () .build (Seq (9, 5, 4, 8, 2, 8, 1, 8, 2, 8, 1, 7, 2 ) )
-    lazy val obtained = input.open (
-      ifEmpty = EmptySeqSD (), ifNonEmpty = sequence => sequence.reverse
+    lazy val obtained = input.open (ifEmpty = EmptySeqSD (), ifNonEmpty = sequence => sequence.reverse
     )
     assert (obtained == expected )
   }