+++ translator/src/test/scala/soda/translator/example/FactorialSpec.scala
@@ -4,8 +4,7 @@ import org.scalatest.funsuite.AnyFunSuite
 
 case class FactorialSpec () extends AnyFunSuite {
 
-  lazy val factorial_values = Seq (
-    (0, 1 ), (1, 1 ), (2, 2 ), (3, 6 ), (4, 24 ), (5, 120 ), (6, 720 ), (7, 5040 ), (8, 40320 ), (9, 362880 ), (10, 3628800 )
+  lazy val factorial_values = Seq ((0, 1 ), (1, 1 ), (2, 2 ), (3, 6 ), (4, 24 ), (5, 120 ), (6, 720 ), (7, 5040 ), (8, 40320 ), (9, 362880 ), (10, 3628800 )
   )
 
   test ("should test the factorial - concise version") {